+++ ThornClient/Modules/Misc/RunInBackground.cs
@@ -1,4 +1,6 @@
+using NukeLib.Utils;
 using ThornClient.Core;
+using ThornClient.Core.ConfigurableElements;
 using ThornClient.Managers;
 using ThornClient.System;
 using UnityEngine;
@@ -14,19 +16,34 @@ public class RunInBackground : Module {
     /// <inheritdoc />
     public override string[] Tags => ["inactive", "afk", "background", "naruto"];
 
+    public Setting<bool> OnlyWhenPaused;
+
     /// <inheritdoc />
     public RunInBackground() : base("thorn.runInBackground", "Run in Background",
         "Keeps the game running when the window is not focused",
         ModuleCategory.Misc) {
+        OnlyWhenPaused = CreateSetting("onlyWhenPaused", "Only when paused/not in game", "Useful for example when waiting for an Angry download", false);
     }
 
     /// <inheritdoc />
     protected override void OnEnable() {
         Application.runInBackground = true;
+        OnlyWhenPaused.OnChanged += UpdateBackgroundRunning;
+        SceneUtils.SafeSceneLoadedNoParam += UpdateBackgroundRunning;
     }
 
     /// <inheritdoc />
     protected override void OnDisable() {
+        OnlyWhenPaused.OnChanged -= UpdateBackgroundRunning;
+        SceneUtils.SafeSceneLoadedNoParam -= UpdateBackgroundRunning;
         Application.runInBackground = false;
+    }
+
+    private bool IsInMenus() {
+        return !SceneUtils.IsInGame() || (OptionsManager.Instance?.paused ?? true);
+    }
+
+    private void UpdateBackgroundRunning() {
+        Application.runInBackground = IsEnabled && (!OnlyWhenPaused.Value || IsInMenus());
     }
 }